Capture the exit code of bash commands and bubble it up
This was surprisingly more complicated than expected. Bash does
not return the exit code for -c commands, so we have to write it
to a file under bash, then read that to get the result under WIN32.
An added complication is that to get full escape support, we needed
to use double quotes for the -c parameter, but using $? within
double quotes leads to substitution of the current $? value (always
0) before the echo, not after; so we have to use another hack to
get this variable to be dynamically created then echoed.